The above code defines a class called `TrialShortNamer`. It has various methods to generate short names for the parameters of an experiment.
`set_defaults` sets the prefix and default values to be used while generating short names for the params.  
`shortname_for_word` generates a short word for the given word.  
`shortname_for_key` generates shortname for the given parameter name.  
`add_new_param_name` adds a new parameter name to the short_param and reverse_short_param list.  
`build_naming_info` generates short name for all the parameter names in the DEFAULTS list.    
`shortname` generates the final short name of the trial using the given parameters and the generated short names.  
`parse_repr` parses the trial short name and returns the experiment's parameters.  

The code has been refactored to follow PEP8 style guide. The code is now properly formatted and it is now easy to read and understand the code. Various comments have been added to the code for better understanding of the code.